# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.CreateTestResultRequest:
    properties:
      comment:
        type: string
      duration_ms:
        type: integer
      status:
        type: string
      test_case_id:
        type: string
    type: object
  entities.CreateTestRunRequest:
    properties:
      build_version:
        type: string
      environment:
        type: string
      name:
        type: string
      plan_id:
        type: string
    type: object
  entities.CreateTestSuiteRequest:
    properties:
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    type: object
...
  entities.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  entities.TestCase:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      expected:
        type: string
      id:
        type: string
      preconditions:
        type: string
      priority:
        type: string
      project_id:
        type: string
      status:
        type: string
      steps:
        type: string
      suite_id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.TestResult:
    properties:
      comment:
        type: string
      created_at:
        type: string
      duration_ms:
        type: integer
      executed_by:
        type: string
      id:
        type: string
      run_id:
        type: string
      status:
        type: string
      test_case_id:
        type: string
    type: object
  entities.TestRun:
    properties:
      build_version:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      environment:
        type: string
      id:
        type: string
      name:
        type: string
      plan_id:
        type: string
      project_id:
        type: string
      status:
        type: string
    type: object
  entities.TestSuite:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      project_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Plain-TCMS API
  version: "1.0"
paths:
...
  /projects/{id}/test-runs:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.TestRun'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test runs by project
      tags:
      - Test Runs
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Run
        in: body
        name: test_run
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestRunRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.TestRun'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Create test run
      tags:
      - Test Runs
  /projects/{id}/test-suites:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.TestSuite'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test suites by project
      tags:
      - Test Suites
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Suite
        in: body
        name: test_suite
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestSuiteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.TestSuite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Create test suite
      tags:
      - Test Suites
...
  /test-runs/{id}:
    delete:
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete test run
      tags:
      - Test Runs
    get:
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestRun'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test run by ID
      tags:
      - Test Runs
    put:
      consumes:
      - application/json
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Run
        in: body
        name: test_run
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestRunRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestRun'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Update test run
      tags:
      - Test Runs
  /test-runs/{id}/cases:
    get:
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.TestCase'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test cases in a run
      tags:
      - Test Runs
    post:
      consumes:
      - application/json
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      - description: Test case ID
        in: body
        name: case
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Add test case to run
      tags:
      - Test Runs
  /test-runs/{id}/cases/{test_case_id}:
    delete:
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Case ID
        in: path
        name: test_case_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Remove test case from run
      tags:
      - Test Runs
  /test-runs/{id}/results:
    get:
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.TestResult'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test results for a run
      tags:
      - Test Runs
    post:
      consumes:
      - application/json
      parameters:
      - description: Test Run ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Result
        in: body
        name: result
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestResultRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.TestResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Add test result to run
      tags:
      - Test Runs
  /test-suites/{id}:
    delete:
      parameters:
      - description: Test Suite ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete test suite
      tags:
      - Test Suites
    get:
      parameters:
      - description: Test Suite ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestSuite'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test suite by ID
      tags:
      - Test Suites
    put:
      consumes:
      - application/json
      parameters:
      - description: Test Suite ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Suite
        in: body
        name: test_suite
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestSuiteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestSuite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Update test suite
      tags:
      - Test Suites
swagger: "2.0"